refactor(providers): split relay & connection providers out of app_providers.dart
Fifth batch of #4506. Moves the relay/connection infrastructure cluster to
`lib/providers/relay_providers.dart`:

- connectionStatusService
- relayCapabilityService
- backgroundActivityManager
- relayStatisticsService
- relayStatisticsStream (Stream<Map<String, RelayStatistics>>)
- relayStatisticsBridge (@keepAlive eager — app-shell-activated)
- relaySetChangeBridge (@keepAlive eager — debounced reconnect+reset)
- _setsEqual<T> file-private helper

`videoFilterBuilder` and `videoVisibilityManager` are deliberately left in
`app_providers.dart` (they sit interleaved between relayCapability and
backgroundActivityManager in the source). They are video-domain providers
and will move with batch 8 (Video & Feed).

`app_providers.dart` adds both `import 'relay_providers.dart';` and
`export 'relay_providers.dart';`. The import is needed because four
remaining providers in `app_providers.dart` still consume
`connectionStatusServiceProvider`, and `videoFilterBuilder` still consumes
`relayCapabilityServiceProvider`. The export keeps all existing consumer
import paths working unchanged.

Removed now-unused imports from `app_providers.dart`:
`background_activity_manager`, `connection_status_service`,
`relay_capability_service`, `relay_statistics_service`. Narrowed the
`nostr_client` import show-list from `NostrClient, RelayConnectionStatus,
RelayState` down to just `NostrClient`.

Codegen moved: 7 Provider chunks relocated from `app_providers.g.dart`
into `relay_providers.g.dart`.

Three follow-up findings (out of scope for this PR, will be filed
separately):
1. Both bridge providers (relayStatisticsBridge and relaySetChangeBridge)
   contain repository-layer state-machine logic embedded in providers —
   same shape as the dirty-sync bridge from batch 4. Candidates for
   extraction into NotifierService/Coordinator classes.
2. backgroundActivityManagerProvider has zero consumers anywhere; the
   underlying singleton is reached via the static factory in
   analytics_service.dart. Joins notificationServiceEnhancedProvider
   from batch 4 in the dead-provider follow-up.
3. relaySetChangeBridge uses Log.error for an expected network failure
   (relay reconnect); per error_handling.md matrix this should be
   Log.warning.

Refs #4506, #4339 Wave 2
