+++ mobile/lib/providers/app_providers.dart
@@ -24,8 +24,7 @@ import 'package:http/http.dart';
 import 'package:keycast_flutter/keycast_flutter.dart';
 import 'package:likes_repository/likes_repository.dart';
 import 'package:models/models.dart' hide LogCategory;
-import 'package:nostr_client/nostr_client.dart'
-    show NostrClient, RelayConnectionStatus, RelayState;
+import 'package:nostr_client/nostr_client.dart' show NostrClient;
 import 'package:nostr_key_manager/nostr_key_manager.dart';
 import 'package:openvine/constants/app_constants.dart';
 import 'package:openvine/extensions/video_event_extensions.dart';
@@ -42,6 +41,7 @@ import 'package:openvine/providers/environment_provider.dart';
 import 'package:openvine/providers/nostr_client_provider.dart';
 import 'package:openvine/providers/og_viner_cache_provider.dart';
 import 'package:openvine/providers/preferences_providers.dart';
+import 'package:openvine/providers/relay_providers.dart';
 import 'package:openvine/providers/saved_sounds_provider.dart';
 import 'package:openvine/providers/shared_preferences_provider.dart';
 import 'package:openvine/services/account_deletion_service.dart';
@@ -50,7 +50,6 @@ import 'package:openvine/services/age_verification_service.dart';
 import 'package:openvine/services/analytics_service.dart';
 import 'package:openvine/services/api_service.dart';
 import 'package:openvine/services/auth_service.dart' hide UserProfile;
-import 'package:openvine/services/background_activity_manager.dart';
 import 'package:openvine/services/badges/badge_repository.dart';
 import 'package:openvine/services/blocklist_content_filter.dart';
 import 'package:openvine/services/bookmark_service.dart';
@@ -62,7 +61,6 @@ import 'package:openvine/services/collaborator_invite_local_state_adapter.dart';
 import 'package:openvine/services/collaborator_invite_service.dart';
 import 'package:openvine/services/collaborator_invite_state_store.dart';
 import 'package:openvine/services/collaborator_response_service.dart';
-import 'package:openvine/services/connection_status_service.dart';
 import 'package:openvine/services/content_deletion_service.dart';
 import 'package:openvine/services/content_filter_service.dart';
 import 'package:openvine/services/content_reporting_service.dart';
@@ -90,8 +88,6 @@ import 'package:openvine/services/pending_action_service.dart';
 import 'package:openvine/services/pending_verification_service.dart';
 import 'package:openvine/services/performance_monitoring_service.dart';
 import 'package:openvine/services/personal_event_cache_service.dart';
-import 'package:openvine/services/relay_capability_service.dart';
-import 'package:openvine/services/relay_statistics_service.dart';
 import 'package:openvine/services/seen_videos_service.dart';
 import 'package:openvine/services/social_service.dart';
 import 'package:openvine/services/subscribed_list_video_cache.dart';
@@ -122,6 +118,7 @@ export 'nostr_apps_providers.dart';
 export 'notifications_providers.dart';
 export 'permissions_providers.dart';
 export 'preferences_providers.dart';
+export 'relay_providers.dart';
 
 part 'app_providers.g.dart';
 
@@ -199,14 +196,6 @@ final badgeRepositoryProvider = Provider<BadgeRepository>((ref) {
 // =============================================================================
 // FOUNDATIONAL SERVICES (No dependencies)
 // =============================================================================
-
-/// Connection status service for monitoring network connectivity
-@Riverpod(keepAlive: true)
-ConnectionStatusService connectionStatusService(Ref ref) {
-  final service = ConnectionStatusService();
-  ref.onDispose(service.dispose);
-  return service;
-}
 
 /// Pending action service for offline sync of social actions
 /// Returns null when not authenticated (no userPubkey available)
@@ -315,14 +304,6 @@ OutgoingDmRetryService? outgoingDmRetryService(Ref ref) {
   return service;
 }
 
-/// Relay capability service for detecting NIP-11 Divine extensions
-@Riverpod(keepAlive: true)
-RelayCapabilityService relayCapabilityService(Ref ref) {
-  final service = RelayCapabilityService();
-  ref.onDispose(service.dispose);
-  return service;
-}
-
 /// Video filter builder for constructing relay-aware filters with server-side sorting
 @riverpod
 VideoFilterBuilder videoFilterBuilder(Ref ref) {
@@ -334,192 +315,6 @@ VideoFilterBuilder videoFilterBuilder(Ref ref) {
 @riverpod
 VideoVisibilityManager videoVisibilityManager(Ref ref) {
   return VideoVisibilityManager();
-}
-
-/// Background activity manager singleton for tracking app foreground/background state
-@Riverpod(keepAlive: true)
-BackgroundActivityManager backgroundActivityManager(Ref ref) {
-  return BackgroundActivityManager();
-}
-
-/// Relay statistics service for tracking per-relay metrics
-@Riverpod(keepAlive: true)
-RelayStatisticsService relayStatisticsService(Ref ref) {
-  final service = RelayStatisticsService();
-  ref.onDispose(service.dispose);
-  return service;
-}
-
-/// Stream provider for reactive relay statistics updates
-/// Use this provider when you need UI to rebuild when statistics change
-@riverpod
-Stream<Map<String, RelayStatistics>> relayStatisticsStream(Ref ref) async* {
-  final service = ref.watch(relayStatisticsServiceProvider);
-
-  // Emit current state immediately
-  yield service.getAllStatistics();
-
-  // Create a stream controller to emit updates on notifyListeners
-  final controller = StreamController<Map<String, RelayStatistics>>();
-
-  void listener() {
-    if (!controller.isClosed) {
-      controller.add(service.getAllStatistics());
-    }
-  }
-
-  service.addListener(listener);
-  ref.onDispose(() {
-    service.removeListener(listener);
-    controller.close();
-  });
-
-  yield* controller.stream;
-}
-
-/// Bridge provider that connects NostrClient relay status updates to
-/// RelayStatisticsService.
-///
-/// Tracks connection/disconnection events via the relay status stream and
-/// periodically syncs per-relay SDK counters (events received, queries sent,
-/// errors) so each relay displays its own real statistics.
-///
-/// Must be watched at app level to activate the bridge.
-@Riverpod(keepAlive: true)
-void relayStatisticsBridge(Ref ref) {
-  final nostrService = ref.watch(nostrServiceProvider);
-  final statsService = ref.watch(relayStatisticsServiceProvider);
-
-  // Track previous states to detect connection changes
-  final Map<String, bool> previousStates = {};
-
-  // Helper to process status updates (used for both initial state and stream)
-  void processStatuses(Map<String, RelayConnectionStatus> statuses) {
-    for (final entry in statuses.entries) {
-      final url = entry.key;
-      final status = entry.value;
-      final wasConnected = previousStates[url] ?? false;
-      final isConnected =
-          status.isConnected || status.state == RelayState.authenticated;
-
-      // Only record changes to avoid excessive updates
-      if (isConnected && !wasConnected) {
-        statsService.recordConnection(url);
-      } else if (!isConnected && wasConnected) {
-        statsService.recordDisconnection(url, reason: status.errorMessage);
-      }
-
-      previousStates[url] = isConnected;
-    }
-
-    // Prune entries for relays no longer in the status map
-    previousStates.removeWhere((url, _) => !statuses.containsKey(url));
-  }
-
-  // Process current state immediately (relays may have connected before
-  // the bridge started)
-  processStatuses(nostrService.relayStatuses);
-
-  // Listen to relay status stream for future connection changes
-  final subscription = nostrService.relayStatusStream.listen(processStatuses);
-
-  // Periodically sync per-relay SDK counters so each relay shows its own
-  // real statistics (not identical values distributed from app-level totals).
-  final syncTimer = Timer.periodic(const Duration(seconds: 3), (_) {
-    final counters = nostrService.getRelayPoolCounters();
-    for (final entry in counters.entries) {
-      statsService.syncSdkCounters(
-        entry.key,
-        eventsReceived: entry.value.eventsReceived,
-        queriesSent: entry.value.queriesSent,
-        errors: entry.value.errors,
-      );
-    }
-  });
-
-  ref.onDispose(() {
-    syncTimer.cancel();
-    subscription.cancel();
-  });
-}
-
-/// Bridge provider that detects when the configured relay set changes
-/// (relays added or removed) and triggers a full feed reset+resubscribe.
-/// Debounces for 2 seconds to collapse rapid add/remove operations.
-/// Only reacts to set membership changes, not connection state flapping.
-@Riverpod(keepAlive: true)
-void relaySetChangeBridge(Ref ref) {
-  final nostrService = ref.watch(nostrServiceProvider);
-  final videoEventService = ref.watch(videoEventServiceProvider);
-
-  Set<String> previousRelaySet = nostrService.relayStatuses.keys.toSet();
-  Timer? debounceTimer;
-
-  void processStatuses(Map<String, RelayConnectionStatus> statuses) {
-    final currentRelaySet = statuses.keys.toSet();
-
-    // Only trigger if the set of relay URLs has changed (not just status)
-    if (!_setsEqual(currentRelaySet, previousRelaySet)) {
-      Log.info(
-        'Relay set changed: '
-        '${previousRelaySet.length} -> ${currentRelaySet.length} relays',
-        name: 'RelaySetChangeBridge',
-        category: LogCategory.relay,
-      );
-
-      previousRelaySet = currentRelaySet;
-
-      // Debounce: collapse rapid changes into a single reset
-      debounceTimer?.cancel();
-      debounceTimer = Timer(const Duration(seconds: 2), () async {
-        Log.info(
-          'Debounce elapsed - forcing WebSocket reconnection and feed reset',
-          name: 'RelaySetChangeBridge',
-          category: LogCategory.relay,
-        );
-
-        // CRITICAL FIX: Force reconnect all WebSocket connections
-        // When relays are added/removed, the existing WebSocket connections
-        // can become stale/zombie - showing as "connected" but not responding
-        // to subscription requests. Force disconnect and reconnect all relays
-        // to establish fresh connections.
-        try {
-          await nostrService.forceReconnectAll();
-          Log.info(
-            'Successfully reconnected all relay WebSockets',
-            name: 'RelaySetChangeBridge',
-            category: LogCategory.relay,
-          );
-        } catch (e) {
-          Log.error(
-            'Failed to reconnect relays: $e',
-            name: 'RelaySetChangeBridge',
-            category: LogCategory.relay,
-          );
-        }
-
-        // Now reset and resubscribe all feeds with fresh connections
-        videoEventService.resetAndResubscribeAll();
-      });
-    }
-  }
-
-  // Process current state immediately to establish baseline
-  processStatuses(nostrService.relayStatuses);
-
-  // Listen to relay status stream for future updates
-  final subscription = nostrService.relayStatusStream.listen(processStatuses);
-
-  ref.onDispose(() {
-    debounceTimer?.cancel();
-    subscription.cancel();
-  });
-}
-
-/// Helper to compare two sets for equality
-bool _setsEqual<T>(Set<T> a, Set<T> b) {
-  if (a.length != b.length) return false;
-  return a.containsAll(b);
 }
 
 /// Analytics service with opt-out support.
